requirements.txt
google-generativeai
python-dotenv
streamlit
tabulate